% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{HydroGen}
\alias{HydroGen}
\title{Hydrography generator}
\usage{
HydroGen(data, by = 10)
}
\arguments{
\item{data}{data frame including environmental parameters.}

\item{by}{numeric, interval between each data point, Default: 10.}
}
\description{
Function generates customized hydrography for test purposes.
}
\details{
Generates a customized hydrography. More details in 'Examples'.
}
\examples{
#Example generates hydro including temperature, salinity and oxygen

data <- data.frame(Temperature = c(17, 15, 10,   8,  4, 4.5,   6,  8),
                   Oxygen =      c(6,   5,  4,   3,  2,   1, 1.5,  0),
                   Salinity =    c(7, 7.5,  8, 8.5,  9, 9.5,  14, 20))
                   
HydroGen(data = data, by = 10)
}
