% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LightBornholm}
\alias{LightBornholm}
\title{Light regime generator}
\usage{
LightBornholm(depth, daytime = seq(0, 24, 0.01666667), ATTk, min, lx)
}
\arguments{
\item{depth}{numeric vector, depth profile which is used by light model, Default: c(1:100).}

\item{daytime}{numeric vector, time sequence in decimals, e.g. every single minute over 24 hours (default).}

\item{ATTk}{numeric, attenuation coefficient.}

\item{min}{numeric, minimum illumination at shallowest depth.}

\item{lx}{logical, if TRUE, output is in lux [lx], if FALSE, computations are in W/m2 (default).}
}
\description{
Function generates semi-customized light profile typically found in the Bornholm Basin (Central Baltic Sea) in early May.
}
\details{
Function makes use of light model developed by (...). Further details in 'Examples'.
}
\examples{
\dontrun{

#The minimum illumination at >0m of depth is assumed to be 2 lx
minVal <- 2

#Attenuation coefficiant is set to -0.16, which can be considered as relatively 'clear'
atteCo <- -0.16

#Computing intra daily under water light regime [lx] down to 100m of depth in Bornholm Basin
testmatrix <- LightBornholm(depth = c(1:100), ATTk = atteCo, lx = TRUE, min = minVal)

#Assigning time of day to each column
ts <- format(seq(from = as.POSIXct("2017-08-19 0:00"), length.out = 1440, by = "min"), "\%H:\%M")
colnames(testmatrix) <- ts

#Printing underwater light regime between 20 and 40 meters of depth at 00:00, 06:00, 12:00 and 18:00
testmatrix[20:40, c("00:00", "06:00", "12:00", "18:00")]
}
}
