% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{parGrad}
\alias{parGrad}
\title{Calculating gradients}
\usage{
parGrad(depth, par, range = 5)
}
\arguments{
\item{depth}{numeric vector of depth indication.}

\item{par}{numeric vector of single environmental parameter (e.g. water temperature).}

\item{range}{numeric integer indicating the range for which the delta should be calculated, Default: 5.}
}
\description{
Function calculates in which direction and at what rate a certain parameter changes over depth.
}
\examples{
#Generating customized temperature and salinity profile
hydro <- HydroGen(data.frame(temp = c(17,10,6,8,9),
                             salt = c(7,8,9,18,20)),
                             by = 25)
                                
#Calculating salinity and temperature gradient on a 3-meter-range
hydro$tempGrad <- parGrad(depth = seq_along(hydro$temp), par = hydro$temp, range = 3)
hydro$saltGrad <- parGrad(depth = seq_along(hydro$salt), par = hydro$salt, range = 3)

head(hydro, 15)
}
