% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SpecificArealOverlap}
\alias{SpecificArealOverlap}
\title{Specific areal overlap}
\usage{
SpecificArealOverlap(data, col = 1)
}
\arguments{
\item{data}{matrix or data frame including two colums - one for each population's distribution.}

\item{col}{numeric integer indicating the column in which the species if interest in given, Default = 1.}
}
\description{
Function calculates fraction of areas occupied by one population on areas occupied by 
another population.
}
\details{
The specific areal overlap is defined as the area occupied by both populations devided by 
the area occupied by one of these populations.
}
\examples{
#Example shows water column occupied by clupeid fish and cod

df <- data.frame(Clupeids = c(0,1,1,1,1,0),
                      Cod = c(0,0,0,1,1,1))

#Calculating specific areal overlap for clupeids and cod, respectively

clu <- SpecificArealOverlap(df, col = 1) * 100
cod <- SpecificArealOverlap(df, col = 2) * 100

paste(clu, "\% of the water layers occupied by clupeid fish are also occupied by cod.", sep = "")
paste(cod, "\% of the water layers occupied by cod are also occupied by clupeid fish.", sep = "")
}
